[move-function] Convert an assert into an early exit.
This assert was making sure we never saw a reborrow since they shouldn't occur
in Raw SIL. Some people were reporting that they /are/ hitting this assert, so I
am converting it into an early exit + skip analyzing an address. This will
prevent the assertion and also will fulfill the same purpose as the original,
not performing the move checking.

This will still result in correctness since if we skip as a move marker
instruction as a result of us skipping processing an address, the pass will emit
a compile time diagnostic saying the checker wasn't able to understand the given
code. So not the best, but at least if this hits a move itself we will be ok.

(cherry picked from commit eb03fb6d6c4615e9f4e47fbf89ddfdd73f78bdd9)