[Driver] Respect -use-ld in DarwinToolChain
The Darwin toolchain is the only toolchain that doesn't use Clang as the
linker driver, and the only one that doesn't respect the value of
-use-ld and completely ignores it. This makes it very complicated to
check Swift with alternative linkers like LLD, zld or mold.

In order to make the behaviour more consistent between toolchains,
introduce a small modification in DarwinToolChains to use the provided
value of `-use-ld`. The newly implement behaviour tries to keep all the
previous behaviour (using the system or the tools directory `ld`), but
opens the door to alternative linkers.

There is small test that checks for some of the new behaviours.