Implement Generic Argument Binding for Variadic Types
The algorithm is detailed below:

Suppose we encounter a type T<..., U, V..., W, ...> for V... the _sole_ variadic generic parameter.

When we encounter an argument list <A, B, C, D, E, F, ...> to T, we must work in three steps

- Bind all generic parameters up to U in parallel with the argument list until we encounter V...
- Measure the tail of parameters after V... and call it `t`. Assuming `m` type variables hav been bound already, we must bind `n - t - m` type arguments to V...
- Finally, bind the remaining `t` arguments.

This procedure is often called "Saturation" in the literature. Variadic generics have a special saturation property where unlike normal generic parameters it is possible to bind zero arguments to them. In such a case, the result is an empty pack type, which is semantically well-formed.