[ShrinkBorrowScope] Checked before adding preds.
Previously, after determining that a block was dead, all its
predecessors were added to the worklist.  That is a problem, with the
current algorithm, if any of the predecessors have a terminator over
which the borrow scope cannot be shrunk.  Here, before adding a dead
block's predecessors to the worklist of blocks through which to shrink,
ensure that all those predecessors have terminators over which the scope
can be shrunk.  Finally, when beginning to process a block, if it
doesn't have a starting instruction, first hoist over thet terminator.