Rip out the concept of an "override job" from default actors.
This has been dead since before the first release.  We need something
like it, but we should probably do it on a clean foundation rather than
building on what's already there.

I've left the barest foundation for "messages" that can be dropped in
the queue but aren't real jobs.

I've left the barest skeleton for "messages" rather than