test: enable SILOptimizer.definite-init-wrongscope
This was matching the wrong `throw` as the scope was incorrect.  Correct
the scope which allows us to match the correct `throw` and enable the
test on Windows.

While here, add an additional check for the `destroy_addr`,
`dealloc_stack` to match the regular exit which is duplicated in the
throwing block.  Add an additional check to ensure that we do not escape
the bounds of the function.

Thanks to @adrian-prantl for helping ensure that the changes here are
valid!

Fixes: SR-14267