[Sema] When explicit existential types are enabled, type witnesses that resolve to ProtocolType or ProtocolCompositionType are always existential types.
This can happen because type aliases to protocol constraints can also be
used as type witnesses. The type alias itself is still a constraint,
because it can be used in generic signatures in addition to being used
as an existential type, e.g. in parameter position. So, we need to wrap
the type alias in an ExistentialType during type witness resolution.