[Named opaque types] Build a generic signature for named opaque types.
For a declaration that has a named opaque type, create a complete
generic signature describing the opaque type that includes the specified
generic parameters and any requirements within that generic parameter
list. This aligns the representation of named opaque types with
unnamed ('some P') opaque result types, so we get the same inference
behavior.

Introduce an egregious hack in type resolution to fake name lookup for
the generic parameters that are specified as part of a named opaque
type. This is to be replaced with the proper name lookup mechanism to
find those generic parameters, such that type resolution can turn them
into opaque type archetypes.

There are also a number of issues with structural opaque types
that are brought to the fore by named opaque types, especially the fact
that OpaqueTypeDecl does not retain the mapping from specific instances
of `some` to the corresponding (implicit) generic parameter, and
generally assumes that the underlying type of the OpaqueTypeDecl itself
is a generic type parameter. Addressing these shortcomings will benefit
both features.